+++ src/main/java/com/leonardo/jikanwapper/records/anime/full/AnimeFull.java
@@ -25,7 +25,7 @@ public record AnimeFull(
         Boolean airing,
         String duration,
         String rating,
-        Double score,
+        Float score,
         @JsonAlias("scored_by")
         Integer scoredBy,
         Integer rank,
